requirements.txt
pandas
matplotlib
reportlab
scikit-learn
python-dotenv
openai